I make some changes

+++ src/Pages/Showproduct.jsx
@@ -1,4 +1,4 @@
-import React, { useEffect } from 'react';
+import React, { useState, useEffect } from 'react';
 import { Link } from 'react-router-dom';
 import { useSelector, useDispatch } from 'react-redux';
 import { experimentalStyled as styled } from '@mui/material/styles';
@@ -14,6 +14,8 @@ import { show } from '../Allreducers/showslice';
 import { deleteproduct } from '../Allreducers/deleteslice'
 import { Button } from '@mui/material';
 import Layout from '../Common/Layout';
+import Swal from 'sweetalert2'; // Import Sweet Alert 
+import { Pagination } from "@mui/material"; // Import Pagination
 
 const Item = styled(Paper)(({ theme }) => ({
     backgroundColor: theme.palette.mode === 'dark' ? '#1A2027' : '#fff',
@@ -25,19 +27,44 @@ const Item = styled(Paper)(({ theme }) => ({
 
 const Showproduct = () => {
     const dispatch = useDispatch();
-    const { showdata, loading } = useSelector((state) => state.Show);
+    const { showdata, totalpage, loading } = useSelector((state) => state.Show);
+    const [totalRecords, setPage] = useState() // Hook For Pagination
 
+    // Handle Make For Pagination 
+    const handleChange = (e, pageno) => { // e is not using but keep as a parameter
+        setPage(pageno);
+        dispatch(show({ page: pageno, perpage: 10 }));
+    };
 
     useEffect(() => {
         dispatch(show());
     }, []);
 
 
-    // Make Handle For Delete
+    // Make Handle For Delete (Start)
     const handleDelete = async (id) => {
-        await dispatch(deleteproduct(id))
-        dispatch(show());
+        // For Sweet Alert
+        const result = await Swal.fire({
+            title: 'Are you sure?',
+            text: 'You will not be able to recover this product!',
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#d33',
+            cancelButtonColor: '#3085d6',
+            confirmButtonText: 'Yes, delete it!'
+        });
+        if (result.isConfirmed) {
+            await dispatch(deleteproduct(id));
+            dispatch(show());
+            // After Deletation Message
+            Swal.fire(
+                'Deleted!',
+                'Your product has been deleted',
+                'success'
+            );
+        }
     }
+    // Make Handle For Delete (End)
 
 
     if (loading) {
@@ -48,62 +75,71 @@ const Showproduct = () => {
         )
     }
 
-
-
     return (
         <>
             <Layout>
-
-                <div className='container mt-5'>
+                <div className='container' style={{ marginTop: '100px' }}>
                     <h1 className='mb-5' style={{ textAlign: 'center' }}>All products</h1>
                     <Box sx={{ flexGrow: 1 }}>
                         <Grid container spacing={2} justifyContent="center">
-                            {showdata.map((value) => (
-                                <Grid item xs={12} sm={6} md={4} lg={3} key={value._id}>
-                                    <Card sx={{
-                                        maxWidth: 345,
-                                        border: '1px solid #E31C25',
-                                        borderRadius: '10px',
-                                        overflow: 'hidden',
-                                    }}>
-                                        <CardActionArea>
-                                            <CardMedia
-                                                component="img"
-                                                height="350"
-                                                image={`https://wtsacademy.dedicateddevelopers.us/uploads/product/${value?.image}`}
-                                                alt="trainer"
-                                            />
-                                            <CardContent>
-                                                <Typography gutterBottom variant="h5" component="div">
-                                                    {value?.title}
-                                                </Typography>
-                                            </CardContent>
-                                        </CardActionArea>
-                                        <CardActionArea>
-
-
-                                            <Button
-                                                onClick={() => handleDelete(value._id)}
-                                            >
-                                                Delete
-                                            </Button>
-
-
-                                            <Link to={`/details/${value._id}`}>
-                                                <Button size="small" color="primary">
-                                                    Details
+                            {showdata.length !== 0 ? (
+                                showdata.map((value) => (
+                                    <Grid item xs={12} sm={6} md={4} lg={3} key={value._id}>
+                                        <Card sx={{
+                                            maxWidth: 345,
+                                            border: '1px solid #E31C25',
+                                            borderRadius: '10px',
+                                            overflow: 'hidden',
+                                        }}>
+                                            <CardActionArea>
+                                                <CardMedia
+                                                    component="img"
+                                                    height="350"
+                                                    image={`https://wtsacademy.dedicateddevelopers.us/uploads/product/${value?.image}`}
+                                                    alt="trainer"
+                                                />
+                                                <CardContent>
+                                                    <Typography gutterBottom variant="h5" component="div">
+                                                        {value?.title}
+                                                    </Typography>
+                                                </CardContent>
+                                            </CardActionArea>
+                                            <CardActionArea>
+                                                <Button onClick={() => handleDelete(value._id)}>
+                                                    Delete
                                                 </Button>
-                                            </Link>
-                                            <Link to={`/updateproduct/${value._id}`}>
-                                                <Button size="small" color="primary">
-                                                    Edit
-                                                </Button>
-                                            </Link>
-                                        </CardActionArea>
-                                    </Card>
-                                </Grid>
-                            ))}
+                                                <Link to={`/details/${value._id}`}>
+                                                    <Button size="small" color="primary">
+                                                        Details
+                                                    </Button>
+                                                </Link>
+                                                <Link to={`/updateproduct/${value._id}`}>
+                                                    <Button size="small" color="primary">
+                                                        Edit
+                                                    </Button>
+                                                </Link>
+                                            </CardActionArea>
+                                        </Card>
+                                    </Grid>
+                                ))
+                            )
+                                :
+                                (
+                                    <>
+                                        <p >No Data Found</p>
+                                    </>
+                                )}
+
                         </Grid>
+                        <Box display="flex" justifyContent="center" alignItems="center" height="100px">
+                            <Grid justifyContent="center">
+                                {showdata.length !== 0 ? (
+                                    <Pagination count={totalpage} onChange={handleChange} totalRecords={totalRecords} />
+                                ) : (
+                                    <></>
+                                )}
+                            </Grid>
+                        </Box>
                     </Box>
                 </div>
             </Layout>
